ui-requirements.txt
streamlit==1.32.0
pandas==2.2.0
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
minio==7.2.0
